LPS-108723 Remove unnecessary Jest mocks
The new version of liferay-npm-scripts sets up some additional useful
mocks by default, allowing us to remove or otherwise simplify the set-up
code in a number of places.
