+++ modules/apps/app-builder/app-builder-web/package.json
@@ -31,9 +31,6 @@
 		"collectCoverageFrom": [
 			"src/main/resources/META-INF/resources/js/**/*.es.js"
 		],
-		"setupFiles": [
-			"<rootDir>/jest-setup.config.js"
-		],
 		"testPathIgnorePatterns": [
 			"<rootDir>/test/js/[^/]+\\.js"
 		]
