+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-web/jest-setup-after-env.config.js
@@ -17,7 +17,3 @@ Liferay.DDM = {
 		spritemap: '/lexicon/icons.svg'
 	}
 };
-
-Liferay.Session = {
-	get: () => 'active'
-};
