+++ modules/apps/frontend-js/frontend-js-web/test/liferay/PortletBase.es.js
@@ -35,12 +35,6 @@ describe('PortletBase', () => {
 	});
 
 	beforeEach(() => {
-		Liferay = {
-			Util: {
-				ns: jest.fn()
-			}
-		};
-
 		portletBase = new PortletBase({
 			namespace
 		});
