+++ modules/apps/segments/segments-web/package.json
@@ -20,9 +20,6 @@
 	},
 	"description": "",
 	"jest": {
-		"setupFiles": [
-			"<rootDir>/jest-setup.config.js"
-		],
 		"testPathIgnorePatterns": [
 			"<rootDir>/test/js/[^/]+\\.js"
 		]
